+++ src/rb/genghis/json.rb
@@ -30,6 +30,7 @@ module Genghis
         when BSON::ObjectId then thunk('ObjectId', o.to_s)
         when BSON::DBRef then db_ref(o)
         when BSON::Binary then thunk('BinData', {'$subtype' => o.subtype, '$binary' => enc_bin_data(o)})
+        when BSON::Timestamp then thunk('Timestamp', {'$t' => o.seconds, '$i' => o.increment})
         when Float then o.nan? ? {'$genghisType' => 'NaN'} : o
         else o
         end
@@ -67,11 +68,12 @@ module Genghis
         when Array then o.map { |e| dec(e) }
         when Hash then
           case o['$genghisType']
-          when 'ObjectId' then mongo_object_id o['$value']
-          when 'ISODate'  then mongo_iso_date  o['$value']
-          when 'RegExp'   then mongo_reg_exp   o['$value']
-          when 'BinData'  then mongo_bin_data  o['$value']
-          when 'NaN'      then Float::NAN
+          when 'ObjectId'  then mongo_object_id o['$value']
+          when 'ISODate'   then mongo_iso_date  o['$value']
+          when 'RegExp'    then mongo_reg_exp   o['$value']
+          when 'BinData'   then mongo_bin_data  o['$value']
+          when 'Timestamp' then mongo_timestamp o['$value']
+          when 'NaN'       then Float::NAN
           else o.merge(o) { |k, v| dec(v) }
           end
         else o
@@ -103,6 +105,9 @@ module Genghis
         BSON::Binary.new(Base64.decode64(value['$binary']), value['$subtype'])
       end
 
+      def mongo_timestamp(value)
+        BSON::Timestamp.new(value['$t'], value['$i'])
+      end
     end
   end
 end